+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 <p align="center"><h2>[Unreleased] (Only in SleepingOwl <code class="language-php">8+</code> branch)</h2></p>
 
+## 2021-02-25
+* [Fix] Fix loading image, images, file, files in `AdminFormElement::hasMany()`: `AdminFormElement::image('image', 'Изображение')`, `AdminFormElement::images('images', 'Изображения')`, `AdminFormElement::file( 'file', 'Файл')`, `AdminFormElement::files('files', 'Файлы')`. В путь добавляется название имени FormElement.
+
 ## 2021-02-06
 * [Fix] Fix `AdminFormElement::image('image', 'Image')->setAssetPrefix($string)` in form.
 * [Update] NPM packages: